api-gateway/requirements.txt
# API Gateway 의존성

# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP 클라이언트
httpx>=0.25.0

# 인증
python-jose[cryptography]>=3.3.0

# 환경 변수
python-dotenv>=1.0.0
